local and youtube functions merged, docker image updated
