+++ Dockerfile
@@ -1,26 +1,23 @@
-FROM ubuntu:22.04
-ARG DEBIAN_FRONTEND=noninteractive
+FROM python:3.10-slim
 
-RUN apt-get update && \
-    apt-get install -y \
-    python3 \
-    python3-pip \
-    python3-dev \
+RUN apt-get update && apt-get install -y \
+    build-essential \
+    software-properties-common \
     ffmpeg \
     && rm -rf /var/lib/apt/lists/*
 
-RUN python3 --version && pip3 --version
-RUN ffmpeg -version
-    
-ENV PYTHONPATH=.
-ENV GRADIO_SERVER_PORT=8501
-ENV GRADIO_SERVER_NAME="0.0.0.0"
-
 WORKDIR /home
 
-COPY requirements.gradio .
+ENV PYTHONPATH=.
 
-RUN pip3 install --no-cache-dir -r requirements.gradio
+COPY requirements.txt /home/requirements.txt
+COPY .streamlit /home/.streamlit
+RUN pip3 install --no-cache-dir -r /home/requirements.txt
 COPY bleep_that_sht /home/bleep_that_sht
+COPY data /home/data
 
-ENTRYPOINT ["python3", "/home/bleep_that_sht/gradio_app_url_download.py"]
+EXPOSE 8501
+
+HEALTHCHECK CMD curl --fail http://localhost:8501/_stcore/health || exit 1
+
+ENTRYPOINT ["streamlit", "run", "/home/bleep_that_sht/app.py", "--server.port=8501", "--server.address=0.0.0.0"]
