--- requirements.gradio
@@ -1,6 +1,0 @@
-whisper-timestamped
-scipy
-moviepy
-yt-dlp
-gradio
-pydub
