--- requirements.streamlit
@@ -1,5 +1,0 @@
-whisper-timestamped
-moviepy
-yt-dlp
-streamlit
-pydub
