+++ requirements.txt
@@ -1,0 +1,90 @@
+aiofiles==23.2.1
+altair==5.5.0
+annotated-types==0.7.0
+anyio==4.6.2.post1
+attrs==24.3.0
+blinker==1.9.0
+cachetools==5.5.1
+certifi==2024.8.30
+charset-normalizer==3.4.0
+click==8.1.7
+cython==3.0.11
+decorator==4.4.2
+dtw-python==1.5.3
+exceptiongroup==1.2.2
+fastapi==0.115.5
+ffmpy==0.4.0
+filelock==3.16.1
+fsspec==2024.10.0
+gitdb==4.0.12
+gitpython==3.1.44
+h11==0.14.0
+httpcore==1.0.7
+httpx==0.27.2
+huggingface-hub==0.26.2
+idna==3.10
+imageio==2.36.0
+imageio-ffmpeg==0.5.1
+jinja2==3.1.4
+jsonschema==4.23.0
+jsonschema-specifications==2024.10.1
+llvmlite==0.43.0
+markdown-it-py==3.0.0
+markupsafe==2.1.5
+mdurl==0.1.2
+more-itertools==10.5.0
+moviepy==1.0.3
+mpmath==1.3.0
+narwhals==1.23.0
+networkx==3.4.2
+numba==0.60.0
+numpy==2.0.2
+openai-whisper==20240930
+orjson==3.10.11
+packaging==24.2
+pandas==2.2.3
+pillow==11.0.0
+proglog==0.1.10
+protobuf==5.29.3
+pyarrow==19.0.0
+pydantic==2.9.2
+pydantic-core==2.23.4
+pydeck==0.9.1
+pydub==0.25.1
+pygments==2.18.0
+python-dateutil==2.9.0.post0
+python-multipart==0.0.12
+pytz==2024.2
+pyyaml==6.0.2
+referencing==0.36.1
+regex==2024.11.6
+requests==2.32.3
+rich==13.9.4
+rpds-py==0.22.3
+ruff==0.7.4
+safehttpx==0.1.1
+scipy==1.14.1
+semantic-version==2.10.0
+setuptools==75.8.0
+shellingham==1.5.4
+six==1.16.0
+smmap==5.0.2
+sniffio==1.3.1
+starlette==0.41.2
+streamlit==1.41.1
+sympy==1.13.1
+tenacity==9.0.0
+tiktoken==0.8.0
+toml==0.10.2
+tomlkit==0.12.0
+torch==2.5.1
+tornado==6.4.2
+tqdm==4.67.0
+typer==0.13.0
+typing-extensions==4.12.2
+tzdata==2024.2
+urllib3==2.2.3
+uvicorn==0.32.0
+websockets==12.0
+whisper-timestamped==1.15.6
+yt-dlp==2024.11.18
